+++ src/main/resources/META-INF/services/org.flowable.cdi.spi.ProcessEngineLookup
@@ -3,4 +3,4 @@
 # SPDX-License-Identifier: EUPL-1.2+
 #
 
-net.atos.zac.flowable.bpmn.ProcessEngineLookupImpl
+net.atos.zac.flowable.processengine.ProcessEngineLookupImpl
